+++ printf.h
@@ -11,7 +11,7 @@ int _putchar(char c);
 int _print_num(int num);
 int _print_string(char *str);
 int _print_unsigned_int(unsigned int num, int base, int flag);
-int _print_address(unsigned long long num);
+int _print_address(unsigned long num);
 int get_len(unsigned int num, int base);
 int _printf(const char *format, ...);
